Creating the layout
First step modeling the layout for the page, setted color variables with the root and main sections of the page created to be filled with content.

Browser reset done with the basics styles of css (margin, padding, html and body)

First thing to improve and I don't understand is the business with before and after properties.