Header Section Completed
In this new commit are included the logo, an input to search and some features to the user navigation.

All setted with flexbox. Some new properties learned.

1. The difference between 'focus' and 'active', and how they work when we interact with them.

But actually I don't really need to use ":focus outline: none" property so it's okey. Hope to use it soon and see how it really works in needed situations.

Finally using the branches and commits in a better way, sorting out each section of the project. Waiting forward to learn Git properly and make big steps learning  big things.