Hotel Details Completed
One more step creating the overview of the hotel.

One new thing that I've never seen before is the mask-image. Used to create list-styles using svg images. Very useful.

Removed the margin from the main container and made it to be all the width, I think is better to use all the available space.

A big question if do I need to keep supporting old browser or fuck them.

No many steps left to finish the first part of the project. :black_joker: