Book Now Completed
The last part of the project completed with a section with user reviews and  another to book a room in the hotel.

Learned how to set before pseudo element with quotes, applying different position.

Added a button with animation and propetiry to hide and appear when hover.

Ready to start with media queries.