Media Queries Completed
All media queries added to the project to be responsive in all the devices.

Actually not many new things in this part but good to remember how to apply them.

A doubt if it everything will work once I upload the repository to Git but for now good job done.

The next changes they'll be to correct all bugs and mistakes already in the browser and phones.